Cleanup of c/c++ code - Removed SDL dependency from VideoState that was introduced by filters - Removed audio filter -- because we don't use it
